extra searching queries for practice and debugging all issues

+++ src/main/resources/schema.sql
@@ -6,36 +6,36 @@ DROP TABLE IF EXISTS roles;
 DROP TABLE IF EXISTS users;
 
 CREATE TABLE users (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255),
     email VARCHAR(255)
 );
 
 CREATE TABLE roles (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255)
 );
 
 CREATE TABLE user_role (
-    user_id BIGINT,
-    role_id BIGINT,
+    user_id INT,
+    role_id INT,
     FOREIGN KEY (user_id) REFERENCES users(id),
     FOREIGN KEY (role_id) REFERENCES roles(id)
 );
 
 CREATE TABLE blogs (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255),
-    manager_id BIGINT,
+    manager_id INT,
     FOREIGN KEY (manager_id) REFERENCES users(id)
 );
 
 CREATE TABLE articles (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(255),
     name TEXT,
-    author_id BIGINT,
-    blog_id BIGINT,
+    author_id INT,
+    blog_id INT,
     FOREIGN KEY (author_id) REFERENCES users(id),
     FOREIGN KEY (blog_id) REFERENCES blogs(id)
 );
